Get webcam working again.
After (a lot of) futzing around with OBS and HTTPS, it seems that OBS' browser plugin doesn't support GetUserMedia, but it does support WebRTC. So, the overlay controller (which runs in a proper browser) grabs the webcam and initializes a VoIP call to the overlay, to send the webcam data across.

Yes, it's complicated, and it's not working perfectly yet, but it's getting there. Apologies to those for whom the https testing affected with associated connections.